azure-data-factory/requirements.txt
datadog
python-dotenv
requests

